Finished first function, still working on 2nd and 3rd, and testcases

+++ requester.py
@@ -1,11 +1,11 @@
-from pandas.io.common import CParserError
 from urllib2 import HTTPError
-from pandas.core.common import PandasError
 import os
 import pandas as pd
 import numpy as np
 import csv
 import requests
+import warnings
+from collections import Counter
 
 
 def get_url(address):
@@ -47,7 +47,8 @@ def url_to_csv(url,name='csv_example'):
         frame.to_csv(name)
         validate_csv(url)
     except Exception:
-        os.remove(name)
+        if os.path.exists(name):
+            os.remove(name)
         raise TypeError('The url entered cannot be parsed into a csv.')
     return os.path.abspath(name)
 
@@ -59,13 +60,15 @@ def batch_url_to_csv(url_list, name_lst):
     """
 
     title_lst = []
-    new_name_lst = []
+    duplicates = [val > 1 for val in Counter(url_list).values()]
+    if True in duplicates:
+        raise AssertionError("Duplicate URLs cannot be present in the parameter 'urls'.")
     for address, name in zip(url_list, name_lst):
-        name = url_to_csv(address, name)
-        new_name_lst.append(name)
-    for f in new_name_lst:
-        title = "{}.csv".format(f)
-        title_lst.append(title)
+        try:
+            name = url_to_csv(address, name)
+            title_lst.append(name)
+        except Exception:
+            warnings.warn("URL was skipped, not a valid url.", RuntimeWarning)
     return title_lst
 
 
@@ -79,7 +82,7 @@ def url_to_df(url):
     try:
         frame = pd.read_csv(url, header=None)
         frame_length_pd = len(frame)
-        frame_length_csv = len(pd.read_csv(url,header=None))#url_to_csv(url)
+        frame_length_csv = len(pd.read_csv(url,header=None))
         if frame_length_pd == frame_length_csv:
             return frame
         else:
@@ -99,80 +102,16 @@ def add(x, y):
 #url_to_csv('https://archive.ics.uci.edu/ml/machine-learning-databazes/car', 'cars')
 #url_to_csv('http://earthquake.usgs.gov/earthquakes/feed/v1.0/summary')
 #url_to_csv('https://www.yahoo.com/')
-CSV_URL = 'https://archive.ics.uci.edu/ml/machine-learning-databases/balloons/yellow-small.data'#'https://archive.ics.uci.edu/ml/machine-learning-databases/car'
-#csv_lst = ['https://archive.ics.uci.edu/ml/machine-learning-databases/balloons/yellow-small.data',
+#fname=url_to_csv(CSV_URL, 'balloons')
+
+
+# #CSV_URL = #'https://archive.ics.uci.edu/ml/machine-learning-databases/balloons/yellow-small.data'#'https://archive.ics.uci.edu/ml/machine-learning-databases/car'
+# #url_to_csv()
+# csv_lst = ['https://archive.ics.uci.edu/ml/machine-learning-databases/balloons/yellow-small.data',
 #             'http://samplecsvs.s3.amazonaws.com/Sacramentorealestatetransactions.csv']
-#file_ex = url_to_csv(CSV_URL, name='sacramento')
-fname=url_to_csv(CSV_URL, 'balloons')
-
-print len(pd.read_csv(fname))
-print len(url_to_df(CSV_URL))
-#validate_csv('/Users/desert/unit_tests_travis/balloons.csv')
-#df = url_to_df(CSV_URL)#, name='sacramento')
-#print df.head()
-
-
-#print(df.head())
-#fs = batch_url_to_csv(csv_lst, ['balloons', 'sacramento'])
-
-
-
-
-# def url_to_csv(CSV_URL=None, name='csv_example'):
-#     with requests.Session() as s:
-#         download = s.get(CSV_URL)
-#         decoded_content = download.content.decode('utf-8')
-#         cr = csv.reader(decoded_content.splitlines(), delimiter=',')
-#         try:
-#             my_list = list(cr)
-#         except UnicodeDecodeError:
-#             yield TypeError
-#         title = "{}.csv".format(name)
-#         outputFile = open(title, mode='w')
-#         outputWriter = csv.writer(outputFile)
-#         #print(len(my_list))
-#         for row in my_list:
-#             outputWriter.writerow(row)
 #
-#
-# Direcory structure: does unittest discovery require __init__ and sub dir?
-#df_rows = url_to_df('http://samplecsvs.s3.amazonaws.com/Sacramentorealestatetransactions.csv')
-#print df_rows
-#
-# def get_url(address):
-#     """Called by url_to_csv, Raises a ValueError for bad URL """
-#     response = requests.get(address)
-#     if response.status_code != 200:
-#         try:
-#         # something
-#         except ValueError as e:
-#
-#             # clean up
-#             raise AssertionError('something went wrong')
-#
-#
-#         else:
-#             pass
-#
-#             # ck csv
-
-
-# def get_url2(address):
-#     """Called by url_to_csv, Raises a ValueError for bad URL """
-#     response = requests.get(address)
-#     if response.status_code != 200:
-#         raise RuntimeWarning
-#     else:
-#         pass
-#
-#
-# def url_to_csv2(url,name='csv_example'):
-#     """Raises a TypeError for urls that cannot be parsed into csv format"""
-#     get_url(url)
-#     try:
-#         frame = pd.read_csv(url, header=None)
-#         return frame
-#     except :
-#         raise RuntimeWarning
-#     except ZeroDivisionError:
-#         raise
+# # csv_lst = ['http://samplecsvs.s3.amazonaws.com/Sacramentorealestatetransactions.csv',
+# #             'http://samplecsvs.s3.amazonaws.com/Sacramentorealestatetransactions.csv']
+# files = batch_url_to_csv(csv_lst,['balloons', 'sacramento'])
+# print files
+# #
